requirements.txt
torch
torchvision
transformers
gradio
pillow
numpy